[WIP] Started MVC-like implementation. Tidying.
