+++ .flaskenv
@@ -1,0 +1,2 @@
+SQLALCHEMY_TRACK_MODIFICATIONS = False
+SQLALCHEMY_DATABASE_URI = "sqlite:///site.db"
