+++ .gitignore
@@ -4,27 +4,11 @@ tweet_data/
 data/
 venv/
 .vscode/
-test/bin
-test/include
-test/lib
-nlp4all/local
-nlp4all/share
-test/share/
-test/local
-nlp4all/bin
-nlp4all/lib
-nlp4all/include
-nlp4all/__pycache__/
-nlp4all/pip-selfcheck.json
-nlp4all/site.db
-flaskblog/.DS_Store
-nlp4all/.DS_Store
-nlp4all/.DS_Store
-*.DS_Store
-*.db
 *.db
 *.xlsx
 *.db
 *.csv
 .ipynb_checkpoints
 .mypy_cache
+.flask_secret
+.env
