+++ nlp4all/__init__.py
@@ -1,27 +1,42 @@
 """
 nlp4all module
 """
-
-import os
+import secrets
+from pathlib import Path
 from flask import Flask
-from flask_sqlalchemy import SQLAlchemy
 from flask_bcrypt import Bcrypt
 from flask_login import LoginManager
-from flask_mail import Mail
+from flask_cors import CORS
+from .models.database import db_session
+from .routes import routes
 
 
-app = Flask(__name__)
-app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
-app.config["SECRET_KEY"] = "5791628bb0b13ce0c676dfde280ba245"
-app.config["SQLALCHEMY_DATABASE_URI"] = "sqlite:///site.db"
-db = SQLAlchemy(app)
+# from flask_mail import Mail
+
+app = Flask(__name__, template_folder="views")
+
+# Load secret key from file, generate if not present
+SECRET_FILE_PATH = Path(".flask_secret")
+try:
+    with SECRET_FILE_PATH.open("r", encoding="utf8") as secret_file:
+        app.secret_key = secret_file.read()
+except FileNotFoundError:
+    # Let's create a cryptographically secure code in that file
+    with SECRET_FILE_PATH.open("w", encoding="utf8") as secret_file:
+        app.secret_key = secrets.token_hex(32)
+        secret_file.write(app.secret_key)
+
+app.session = db_session
+
+app.register_blueprint(routes)
+
+CORS(app)
+
+# db = SQLAlchemy(app)
 bcrypt = Bcrypt(app)
 login_manager = LoginManager(app)
-login_manager.login_view = "login"
+# login_manager.login_view = "login"
 login_manager.login_message_category = "info"
-app.config["MAIL_SERVER"] = "smtp.googlemail.com"
-app.config["MAIL_PORT"] = 587
-app.config["MAIL_USE_TLS"] = True
-app.config["MAIL_USERNAME"] = os.environ.get("EMAIL_USER")
-app.config["MAIL_PASSWORD"] = os.environ.get("EMAIL_PASS")
-mail = Mail(app)
+
+
+# mail = Mail(app)
