+++ nlp4all/models/__init__.py
@@ -1,0 +1,4 @@
+"""Database and models for NLP4ALL."""
+
+from .BayesianRobot import BayesianRobot
+from .BayesianAnalysis import BayesianAnalysis
