+++ nlp4all/routes/__init__.py
@@ -1,0 +1,4 @@
+"""Flask Routes"""
+
+from flask import Blueprint
+routes = Blueprint('routes', __name__)
