+++ nlp4all/routes/auth.py
@@ -1,0 +1,2 @@
+"""Backend routes"""
+
