+++ nlp4all/routes/user.py
@@ -1,0 +1,1 @@
+"""Frontend routes."""
