+++ requirements.txt
@@ -13,13 +13,13 @@ click==8.1.3
 colorclass==2.2.2
 dill==0.3.5.1
 distlib==0.3.6
-Django==4.1.2
 dnspython==2.2.1
 docopt==0.6.2
 email-validator==1.3.0
 filelock==3.8.0
 Flask==2.2.2
 Flask-Bcrypt==1.0.1
+flask_cors==3.0.10
 Flask-Email==1.4.4
 Flask-Login==0.6.2
 Flask-Mail==0.9.1
@@ -60,6 +60,7 @@ pylint-flask-sqlalchemy==0.2.0
 pylint-plugin-utils==0.7
 PySocks==1.7.1
 pytest==7.1.3
+python-dotenv==0.21.0
 pytz==2022.4
 PyYAML==6.0
 requests==2.28.1
